Refactor del codigo y agrega archivo de postman
