+++ src/main/java/com/api/vet/model/Client.java
@@ -1,35 +1,33 @@
 package com.api.vet.model;
 
-import com.fasterxml.jackson.annotation.JsonIgnore;
+
 import com.fasterxml.jackson.annotation.JsonManagedReference;
 import lombok.AllArgsConstructor;
-import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
 import javax.persistence.*;
 import java.util.Date;
 import java.util.HashSet;
-import java.util.List;
 import java.util.Set;
 
 
 @Entity
-@Table(name="CLIENT")
+@Table(name = "CLIENT")
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
 
 public class Client {
     @Id
-    @Column(name="DOCUMENT")
+    @Column(name = "DOCUMENT")
     private Long id;
-    @Column(length = 50, nullable = false, name="ADDRESS")
+    @Column(length = 50, nullable = false, name = "ADDRESS")
     private String address;
-    @Column(name="NAME", nullable = false)
+    @Column(name = "NAME", nullable = false)
     private String name;
 
-    @Column(name="DATE_CREATED")
+    @Column(name = "DATE_CREATED")
     private Date dateCreated;
 
     @JsonManagedReference
@@ -38,14 +36,6 @@ public class Client {
             fetch = FetchType.LAZY,
             cascade = CascadeType.ALL)
     private Set<Pet> pets = new HashSet<>();
-
-
-
-
-
-
-
-
 
 }
