+++ src/main/java/com/api/vet/model/Reservation.java
@@ -1,16 +1,12 @@
 package com.api.vet.model;
 
-import com.fasterxml.jackson.annotation.JsonBackReference;
-import com.sun.istack.NotNull;
+
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
 import javax.persistence.*;
-import java.time.LocalDate;
 import java.util.Date;
-import java.util.HashSet;
-import java.util.Set;
 
 
 @Entity
@@ -22,10 +18,10 @@ public class Reservation {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name="ID")
+    @Column(name = "ID")
     private Long idReserva;
 
-    @Column(nullable = false,name="DATE" )
+    @Column(nullable = false, name = "DATE")
     private Date reservationDate;
 
     @ManyToOne(fetch = FetchType.LAZY)
