+++ src/main/java/com/api/vet/repository/ReservationRepository.java
@@ -1,17 +1,12 @@
 package com.api.vet.repository;
 
 import com.api.vet.model.Client;
-import com.api.vet.model.Pet;
+
 import com.api.vet.model.Reservation;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Query;
-import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
-import org.springframework.web.bind.annotation.PathVariable;
 
 import java.util.Date;
-import java.util.List;
-import java.util.Optional;
 
 @Repository
 public interface ReservationRepository extends JpaRepository<Reservation, Long> {
