+++ src/main/java/com/api/vet/services/pet/ServicePetImp.java
@@ -1,20 +1,18 @@
-package com.api.vet.services;
+package com.api.vet.services.pet;
 
-import com.api.vet.model.Client;
 import com.api.vet.model.Pet;
-import com.api.vet.repository.ClientRepository;
 import com.api.vet.repository.PetRepository;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
-import java.util.Optional;
 
 @Service
 @AllArgsConstructor
-public class ServicePetImp implements ServicePet{
+public class ServicePetImp implements ServicePet {
 
     private final PetRepository repositoryPet;
+
     @Override
     public Pet savePet(Pet pet) {
         Long clientId = pet.getClient().getId();
