+++ src/main/java/com/api/vet/services/reservation/ServiceReservation.java
@@ -1,0 +1,9 @@
+package com.api.vet.services.reservation;
+
+import com.api.vet.model.Reservation;
+
+
+public interface ServiceReservation {
+    Reservation saveReservation(Reservation reservation);
+
+}
